pyproject.toml
[project]
name = "bridget"
version = "1.0.0"
description = "Bot for the palera1n Discord server"
authors = [
    {name = "Nebula", email = "me@itsnebula.net"},
    {name = "June Park", email = "rjp2030@outlook.com"}
]
dependencies = [
    "mongoengine>=0.27.0",
    "discord-py @ git+https://github.com/Rapptz/discord.py",
    "Pillow>=9.5.0",
    "pycryptodome>=3.17",
    "chatgpt @ git+https://github.com/0xallie/chatgpt",
    "APScheduler>=3.10.1",
    "emoji-country-flag>=1.3.2",
    "humanize>=4.6.0",
    "mypy @ git+https://github.com/python/mypy",
    "typeguard>=3.0.2",
    "pyannotate>=1.2.0",
    "retype>=22.8.0",
    "pyre-check>=0.9.18",
]
requires-python = ">=3.10,<4.0"
license = {text = "BSD-3-Clause"}
readme = "README.md"

[project.urls]

[tool.pdm.build]
includes = []

[tool.pdm.scripts]
bot.cmd = "python bridget"
bot.env_file = ".env"
setup.cmd = "python bridget/setup.py"
setup.env_file = ".env"
lint.cmd = "flake8 --config=.flake8"
gentypes.cmd = "python -m mypy.stubs bridget"
applytypes.cmd = "python -m retype -p out bridget -t bridget"

[tool.pyright]
extraPaths = ["__pypackages__/3.11/lib/"]

[tool.black]
line-length = 120
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
experimental-string-processing = true
skip-string-normalization = true
extra = """
    # Add a blank line before a new definition
    def: +2
"""



[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
